refactor(bazel-module): Reorganize parser into the own directory
