+++ lib/modules/manager/bazel-module/parser/module.ts
@@ -1,8 +1,7 @@
-import { lang, query as q } from 'good-enough-parser';
-import { regEx } from '../../../util/regex';
-import { Ctx } from './context';
-import type { RecordFragment } from './fragments';
-import * as starlark from './starlark';
+import { query as q } from 'good-enough-parser';
+import { regEx } from '../../../../util/regex';
+import type { Ctx } from '../context';
+import * as starlark from '../starlark';
 
 const booleanValuesRegex = regEx(`^${starlark.booleanStringValues.join('|')}$`);
 const supportedRules = [
@@ -26,7 +25,7 @@ const kvParams = q
     q.sym<Ctx>(booleanValuesRegex, (ctx, token) => ctx.addBoolean(token.value)),
   );
 
-const moduleRules = q
+export const moduleRules = q
   .sym<Ctx>(supportedRulesRegex, (ctx, token) => ctx.startRule(token.value))
   .join(
     q.tree({
@@ -36,18 +35,3 @@ const moduleRules = q
       postHandler: (ctx, tree) => ctx.endRule(),
     }),
   );
-
-const rule = q.alt<Ctx>(moduleRules);
-
-const query = q.tree<Ctx>({
-  type: 'root-tree',
-  maxDepth: 16,
-  search: rule,
-});
-
-const starlarkLang = lang.createLang('starlark');
-
-export function parse(input: string): RecordFragment[] {
-  const parsedResult = starlarkLang.query(input, query, new Ctx());
-  return parsedResult?.results ?? [];
-}
